Update to latest liblc3 from google repo

+++ v3/attdet.h
@@ -1,6 +1,6 @@
 /******************************************************************************
  *
- *  Copyright 2021 Google, Inc.
+ *  Copyright 2022 Google LLC
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -15,13 +15,6 @@
  *  limitations under the License.
  *
  ******************************************************************************/
-
-/**
- * LC3 - Time domain attack detector
- *
- * Reference : Low Complexity Communication Codec (LC3)
- *             Bluetooth Specification v1.0
- */
 
 #ifndef __LC3_ATTDET_H
 #define __LC3_ATTDET_H
@@ -38,7 +31,7 @@
  * return          1: Attack detected  0: Otherwise
  */
 bool lc3_attdet_run(enum lc3_dt dt, enum lc3_srate sr,
-    int nbytes, lc3_attdet_analysis_t *attdet, const float *x);
+    int nbytes, lc3_attdet_analysis_t *attdet, const int16_t *x);
 
 
 #endif /* __LC3_ATTDET_H */
